Add Rake support to project
In order to provide a set of simple commands for executing the tests we have implemented Rake to allow us to create tasks which know how to run the tests in a Chrome web browser for example.

This will also give us the means to expand the commands or tools we can run as the project expands.