Remove use of RUN_IN_BROWSER env var from project
Prior to this change we required the env var RUN_IN_BROWSER to be set if a non-headless mode was required. We defaulted to firefox if this was set to 'true' but allowed the option of switching to chrome if another env var called BROWSER is set.

This was felt to be overly complex so this change removes the env var RUN_IN_BROWSER. Now the default is to use the headless browser poltergeist, else users should choose a headless browser by setting the env var BROWSER to either chrome or firefox.