Fix web demo tests failing in poltergeist
It was found that though the tests ran using selenium, in poltergeist they were failing. The problem was due to how the google homepage was caputuring the input (it would remove the search box and switch to using its auto filtering alternative) and poltergeist's lack of support for send_keys.

The solution was to redo the tests using a search where the search field and search button remain in place as you enter your text. This change updates the tests to use wikipedia instead. It also adds the launchy gem to help with debugging in future.