Implement site-prism page object model DSL
Experience has shown that keeping our tests dry helps make them more maintainable over time. A critical aspect in doing this is the use of the page object pattern.

A page object wraps up all functionality for describing and interacting with a page into a single object. This object can then be referred to in the steps, rather than risk duplicating the logic in different steps.

This change adds the gem site_prism to the project along with a basic example of how to create a page and steps that use it.