Add @demo custom tag to example features
The intention is that users would either clone or fork this template and use it to start building there own tests. Once they are comfortable with their test suite and how to write them the demo features included here should be deleted to avoid confusion.

However just in case this change adds the tag @demo to the two example feature tests. It also adds a config/cucumber.yml file so we can create profiles and ensure when cucumber is run by default these demo tests are not included.