Rename env var BROWSER to DRIVER to fix Launchy
In testing found an incompatibility with setting an environment variable called BROWSER and trying to use Launchy. It seems this is a variable used by the OS for setting what the default browser is, and in turn Launchy references it when trying to open the saved page.

So this change updates the ENV VAR and corresponding code and comments to use DRIVER instead of BROWSER.