Structure project 2 avoid conflicts with upstream
The intention is to recommend users of Quke fork the project rather than simply clone it. If they do this it means any fixes or enhancements we make to the parent project can still be merged in to their project using an upstream pull.

To ensure this goes as smoothly as possible for the end user this change restructures the project to ensure all files relating to the quke demo features are placed in `quke` folders to keep them separate from those a user creates.

It also includes a renaming of the demo tags and demo page object to include the word 'quke' to avoid any naming conflicts.