Fix use of BROWSER instead of DRIVER in rake
Recently we switched using BROWSER for the name of an environment variable (see https://github.com/EnvironmentAgency/quke/commit/20c12859ccb2fdd35fd746648dba53c53ccc3878) to DRIVER in order to resolve a problem preventing Launchy from working.

However it looks like we failed to update the use of the environment variable in `Rakefile` so this change rectifies that oversight.