Make use of hooks clearer in project
Cucumber has the concept of hooks. These are a way of running code before, after, or around a feature, or even before and after an individual step. We use an `after_step` hook to enable the pause functionality to work.

This change makes the use of hooks clearer in general by creating a folder called `hooks`, and moving the `after_step` hook for pausing between steps to a separate file within the folder.