Add initial .rubocop.yml config to project
Developers within our teams at the EA generally use rubocop to ensure the code they are writing is consistent with community conventions. In order to capture the values of quke's custom arguments for use elsewhere we have saved them to global variables which rubocop raises an excption for.

We know that global variables are not recommended hence rubocop flagging them, but in this case and due to the intended audience of quke (people who are not necessarily developers) it was felt to be the best solution. So this change adds a `.rubocop.yml` file with some default settings (display cop names and style guide link in errors) and GlobalVars disabled.