Add logic to stop too many pages being opened
Experience has shown that should a major element of your service go down all your tests will start failing which means you can be swamped with output from `save_and_open_page`.

To prevent this the change in this commit adds logic to the `after` hook to keep a global count of the number of fails, and if it hits 5 it will cause cucumber to close.