Working version of first 2 pages of FRAE journey
In this working version we now use an App class to bind together the page objects and remove noise from the steps.

Rather than specifying radio buttons and checkboxes individually in the page objects we using SitePrism's `elements` feature to find and manage them.

We have also updated the config to ensure SitePrism uses Capybara's implicit wait system so we don't have to manually cater for this in the steps (though this should be set in quke itself).

Bringing these changes together we now have the first couple of pages of the FRAE journey working.