Make url endpoint to be tested configurable
https://github.com/EnvironmentAgency/quke/issues/8

Currently Quke and its examples have been built on the premise that the website undertest would be at a fixed address set within the test code. However initial feedback has shown this may not always be the case, for example QA may want to test against its own QA environment, then pre-prod, and then production upon release of the service.

This change will add the ability for the user to configure the url using an environment variable.

(Resolves issue #8)