Add embedded test website
Prior to this change the web based tests used wikipedia. This means Quke has a dependency on an external system it does not control. The struture and content could change and it may not always be available.

This change adds an embedded web application in Quke against which the demo tests can then be run. This benefits Quke by

- removing the external dependency
- simplifying CI (something we'd like to add to Quke)
- allowing for more in depth examples of tests and the Site Prism page objects

The last is an important driver for this change. Initial use of Quke has demonstrated that the included examples are not representative or detailed enough for our intended audience. Having an embedded web application which we control means we can generate the pages and content needed to best demonstrate the features of Capybara and SitePrism.