Add 'search' example feature
In order to start demonstrating how to use Quke to test web pages this change adds the first example, which is a search page that displays 2 results if 'capybara' is entered.

This change encompasses the work to add the search functionality to Quke's internal web app, and the accompanying tests.