Add .htmlhintrc file to project
With the inclusion of essentially html pages (the erb views) in the project due to the demo web app we still wanted to ensure the html in those views is valid.

For this we have adopted htmlhint (in some teams). Because the erb files includ embedded code (normally within <% %> tags) we need the config file to ensure these elements are ignored, and to keep consistent with how actual services are checking their html files.

Therefore this change adds a .htmlhintrc file with the same settings as the Waste Exemptions digital service and the Flood Risk Activity Exemptions digital service.