Update rake task 'run' to use rerun if its available
[Rerun](https://github.com/alexch/rerun) is a gem that 'Restarts an app when the filesystem changes'. When working on a Sinatra app like the demo app included this project it can save a developer lots of time as they don't have to keep stopping and starting the app after each code change.

However its not a requirement for Quke to work so have chosen not include it as a dependency in the Gemfile. This change simply updates the rake task 'run' to use a bit of intelligence to work out if its installed and if it is, to start the app using it.