Tidy up of the code
We spotted after creating the new internal app that we had not left the code in a 'clean' state. There are some comments that either need to be moved or removed, and some of the tests such as the non-web ones no longer apply (with no external dependency anymore we don't need these to confirm everything is working).

This change is essentially a bit of house keeping with no real impact on actual functionality.