Stop saving to root when save_and_open_page used
Currently if `save_and_open_page` is called or a page errors when using the selenium-webdriver the current page is saved by Capybara and launched by Launchy automatically.

The problem is the files are being created in the root of the project. The `.gitignore` file is already set to ignore them, but when working on a project they start getting in the way of trying to navigate the source files.

This change sets the Capybara config to save them to a `tmp/` folder instead, and adds a new rake task to allow users to quickly delete them.