Add missing @quke tag to css_selectors feature
When adding new example feature tests we don't add the tag to aid with development. However before finishing and merging into master the tag needs to be added else it will be included in users tests.

However the last feature added, css_selectors omitted the tag so this change fixes that.