Add support for a yaml (.yml) config file
Though currently not really needed we are looking at adding some abilities later, specifically integration with [Browserstack](https://www.browserstack.com) which will require the ability to set many more options and values.

For a user to manage this in environment variables alone was felt unwieldy, so this change adds support for specifying configuration in a `.yml` file as well as retaining support for environment variables.

Subsequent features can then build on this change to add more flexibility, by allowing users to specify options for Quke to use or pass on in the config file.