Move poltergeist driver options to Quke::Config
Rather than instantiating an options hash in `features/support/env.rb` which is then used in the call to `Capybara::Poltergeist::Driver.new(app, options)`, we have moved this to lib/quke/config.rb.

This will form the basis for the approach of passing options to drivers going forward (in preparation for browserstack integration), and gives us the flexibility in future to open these options up to users to configure.