Add support for multiple config filespty message aborts the commit
The initial implementation of support for config files relied on there being only 1 and it being called **config.yml**.

This change extends the implementation to allow support for multiple config files to be created and used. Now **.config.yml** is expected to be the default, but users can override this by setting an environment variable (`QCONFIG`) with the name of the config file to use.

This change also sees in the README a move away from offering the option to use either environment variables or config files, to just using config files. The 3 options can still be set using environment variables, but as Quke is developed further maintaining support for specifying all options in env vars won't be feasible.