Clear up references to poltergeist and phantomjs
In Quke the Selenium web driver is used driver either Firefox or Chrome, and Poltergeist drives the headless browser PhantomJS. However either in code or the instructions we reference 'firefox' and 'chrome' as browser options (not selenium), along with 'poltergeist'. To keep the naming consistent we should be using 'phantomjs' so this change updates references both in code and documentation to use 'phantomjs' (where appropriate).