Fix Cucumber not loading lib/features correctly
Because Cucumber is called in the context of the executing script (project) and not the gem, when it is telling Cucumber to reference `lib/features` Cucumber cannot see it. This means `env.rb` never gets loaded and therefore everything blows up.

This change corrects how we specify our default arguments, ensuring we correctly determine what the full path to the gems `lib/features` folder is, and passing that to Cucumber.