Change after_hook error to use config not assume
This will resolve issue #41, which highlights that **Quke's** assumption that CI is only being when the **phantomjs** driver is selected is restrictive and incorrect.

Users may wish to use other drivers in a CI environment, for example [XVFB](https://en.wikipedia.org/wiki/Xvfb) with Chrome. Therefore this change removes the assumption in the `after_hook.rb` logic to fail immediately if **phantomjs** is being used, to instead drive the behaviour of a config flag.