Add support for proxy servers
This change adds support to **Quke** to specify a proxy server when using the **PhantomJS**, **Firefox**, and **Chrome** drivers.

The reasoning is that a connection to the internet may be been setup to go via a proxy server for security purposes in some build environments. With no way currently to tell **Quke**, it would not be able to access the site under test.

To add this support the following changes have been made

- `Quke::Configuration` will now accept a new section in the `.config.yml` called `proxy`, containing the values `host` and `port`
- `Quke::Configuration` has a new helper method which returns whether proxy settings need to be applied to the driver
- `Quke::DriverRegistration` now contains logic specific to each driver for specifying proxy details (if required)
- The methods which returned config options for **Poltergeist** and **PhantomJS** have been moved to `Quke::DriverRegistration` (more details below)

Though it was envisaged that users could set the options for each driver by simply adding some key value pairs into the `.config.yml`, adding support for using a proxy has shown in practise this is not always the case. For example we need to instantiate a `Selenium::WebDriver::Firefox::Profile.new` in order to support setting the proxy for the Firefox driver.

Therefore these place holders for passing in options have been removed, and the driver configuration for poltergeist and phantomjs has been moved out of `Quke::Configuration` into `Quke::DriverRegistration`. This brings it inline with the other drivers and now means `Quke::Configuration` is responsible solely for what is (or can be) specified in the config yml file.