Refactor driver configuration to new class
Completing the changes for proxy server support ( 62d33f9 ) meant adding more configuration logic to the existing `Quke::DriverRegistration` class. This put it over the line of being **monolithic**, and felt like we were mixing two different domains; the driver registration logic, and the logic for determining each driver's configuration settings.

So this change moves the driver configuration logic into a new class called `Quke::DriverConfiguration`. This has led to knock on changes of how we initialise each of the objects, and how each driver is registered.

The benefits are that both classes are now much easier to maintain, and more of our code has test coverage (something that helped spot some omissions in setting up the **Browserstack** capabilities).