Add support 4 setting 'no proxy' arg. in drivers
Having previously thought we had done enough to support using **Quke** in an environment where a proxy server is in use (PR #47), we found that is not the case. The problem is that a common configuration is to only use the proxy server when making external connections, but not for local ones.

So without the ability to allow **Quke** to configure the drivers in this was, when the **selenium** based drivers attempt to make local connections to the selenium server **Quke** falls over.

This change will add support for specifying this option. In the same manner as PR #47 the user simply has to provide the details in the config file, **Quke** will handle how to configure the drivers with this information when registering them with [Capybara](https://github.com/teamcapybara/capybara).