Add ability to set user-agent to use
https://github.com/DEFRA/quke/issues/58

This change allows users to specify the user agent a driver should use via the `config.yml`. This will only apply to the **Firefox**, **Chrome** and **PhantomJS** drivers.

The changes involved taking the argument passed in and correctly configuring the underlying driver to use it. In the case of **Firefox** this involved updating the profile. For **Chrome** it was simply adding an additional flag.

**PhantomJS** proved to be more of a problem. You have to add a header via **Poltergiest**, however it only last for the current session i.e test. This means before every scenario you have to keep adding it back in.

This meant a new `before()` hook was added to handle this.

Other changes include tests and documentation.