More flexibility setting browserstack capabilities
Prior to this change **Quke** looked for specific values in the `.config.yml` file and then set those as capabilities in the Browserstack driver. However we've had a coup0le of requests for specific things to be passed through. Continuing to support Browserstack in this way would mean continually having to dive back into the code and adding support for each and every flag.

So instead this updates **Quke** to be more flexible. Now it will simply passing in the values set in `.config.yml` (under browserstack) to the `Selenium::WebDriver::Remote::Capabilities` object we create, using the keys in the config as the names of the keys in the capabilities.

Essentially we're moving to an assumption that the user knows what they are doing when it come to Browserstack. So we'll handle setting up the driver and **Quke**, but get out of their way when it comes to browserstack.

It should be noted this is a breaking change for Browserstack users. Both the username and auth_key, and the capabilities to be set for the test used to be found directly under the node `browserstack:`. Now Quke will expect them to be placed under `browserstack: capabilities:` for example

```yaml
browserstack:
  username: jdoe
  auth_key: 123456789ABCDE
  capabilities:
    project: 'Adding browserstack support'
    build: 'Version 1'
```