Add support for browserstack local testing (#69)
[Browserstack local testing](https://www.browserstack.com/local-testing) allows you to test private and internal servers, alongside public URL's. It does this by establishing a secure connection between your machine and BrowserStack servers.

We had previously attempted to support this when Quke was first put together however could not get it to work. At a recent hack day for Quke we were able to get local testing working using the latest method from Browserstack.

There is a binary which must be running when your test starts, and you simply need to set the Browserstack capability flag `browserstack.local` to `true`. For this change we're taking it a step further by having Quke install the binary as a dependency, and then manage starting and stopping it when required.

This keeps with our philosophy of Quke being a framework that makes acceptance testing simple.

This change seems quite large in proportion to the actual feature being added. This is because as we began adding methods to determine things like whether local testing was requested, and the args required to start the binary, we blew the size limits on Quke::Configuration. So it also includes a series of refactors, splitting anything related to configuring Browserstack into a separate configuration class.