Add ability to update browserstack session status (#72)
This change builds on our increased support for [Browserstack](https://www.browserstack.com). Each test run is recorded as a session in **BrowserStack**, and a session has a unique ID and a status. If you know the session ID, you can use **Browserstack's** REST API to set the status to 'passed' or 'failed' (the default is 'Completed').

With this change Quke will update the status for users automatically. It only applies when users have selected the `browserstack` driver, but if they have Quke will mark the session as 'passed' if all tests pass, else 'failed'.

It also includes some housekeeping (typos, documentation improvements, and some tweaks to our rubocop rules).