Update selenium and add geckodriver
We needed to update selenium as Quke was stuck only being able to support testing of Firefox v47 or less due to the change to using Geckodriver in Firefox 48.

When this first happened we help off updating Quke as selenium-webdriver did not have full support yet for Geckodriver.

Since then though those issue have been supported and the latest Selenium Webdriver supports the latest Firefox versions.

The one change is, like chrome an additional driver needs to be installed to drive Firefox. Hence the addition of `geckodriver-helper` which is a project that mirrors the way `chromedriver-helper` works to install and manage the geckodriver binary for you, rather than having to download and start it separately outside your tests.