Update project dependencies (#80)
It has been a long time since Quke's dependencies were updated.

These changes are as a result of updating all our dependencies to the latest and greatest possible. Key updates were

 - development dependencies - All dependencies have been updated as well as some for the version specifiers being removed. Generally we want the latest and greatest when it comes to these so don't need to be cautious and restrict the versions.
 - rspec - Mainly the expectations which Quke uses to support using asserts in tests.
 - site_prism - SitePrism has gone 3.0 since our last updates to Quke. So we have updated to the latest and made some minor changes around config and tags based on its update notes.
 - chromedriver helper - As well as updating the gem the latest docs say you also need to `require` it in your app to prevent `unable to connect to chromedriver 127.0.0.1:9515 (Selenium::WebDriver::Error::WebDriverError)` errors (which we got after just updating the gem). It also instructs you to require `selenium-webdriver` before `chromedriver` so we added that in as well.
 - Poltergeist
 - Capybara

Finally we also managed to update selenium-webdriver. We needed to update selenium as Quke was stuck only being able to support testing of Firefox v47 or less due to the change to using Geckodriver in Firefox 48.

When this first happened we help off updating Quke as selenium-webdriver did not have full support yet for Geckodriver.

Since then though those issue have been supported and the latest Selenium Webdriver supports the latest Firefox versions.

The one change is, like chrome an additional driver needs to be installed to drive Firefox. Hence the addition of `geckodriver-helper` which is a project that mirrors the way `chromedriver-helper` works to install and manage the geckodriver binary for you, rather than having to download and start it separately outside your tests.