Update Selenium to use Options
Since the update to selenium we have been getting deprecation warnings about the use of the `switches:` argument when registering the chrome driver, and the `profile:` argument when registering Firefox.

These have both been deprecated in favour of using `options:`.

So this change updates Quke to use the new way of passing our configuration to the driver when registering it via Capybara.