Add support for headless chrome and firefox
Now Quke has been updated and is using the latest Selenium Webdriver, we can take advantage of the fact they both have a headless mode.

This change adds the ability to tell Quke you wish to use either Chrome or Firefox in their headless mode.

As per phantomjs, we have followed the same behaviour of not launching a view of the error page in the event of an error so the experience is truly headless.