Add support for running tests in parallel (#83)
Because acceptance tests need to run against the actual service using the browser they can be much slower than unit tests.

To help this change adds support for running tests in parallel by implementing the [Parallel Tests gem](https://github.com/grosser/parallel_tests).

Parallel Tests is a bit like Quke in that it takes the arguments passed in and then uses them to call multiple instances of cucumber on multiple threads.

Because if this we are moving away from calling Cucumber through its CLI, and instead using the `ParallelTests::CLI` instead. The difference is if the config option `parallel:` is set
to false, we run ParallelTests in single mode which just starts one process mimicking what would     happen if we called cucumber directly. If set to true, we let ParallelTests do its thing.

Also because `lib/features/support/env.rb` is instantiated separately in each process, it does not have access to the global `Quke::Configuration` hence we now need to instantiate it within `env.rb`. Because of this change and the need to access it within `Quke::CukeRunner` we've also dropped setting the config in the exe file, and instead do it in the `Quke::CukeRunner.initialize`.

N.B. Because of the change to use ParallelTests for all executions, we now get an error if we run with `features_folder` pointing to somewhere that doesn't exist. We were not able to fix the unit test at this time so we're taking the test coverage hit and will come back to it later.