Add support for additional parallel configuration (#85)
Initial feedback for the parallel mode was that depedending on the project under test how features were split across them could be quite unbalanced.

For example in a project with some features containing only one scenario, and others containing lots you could end up waiting for one process to finish because the others got the features with only a few scenarios. So the ability to tell ParallelTests to group by 'scenarios' would be really beneficial.

The other was that depending on your machine you may get some benefit increasing the number of processes running.

Hence this change expands the parallel config from a simple `parallel: true` to a whole new section in the `config.yml`, and addss support into Quke to process the additional arguments.