Add option to control display failures
By default Quke will display web pages where a failure has taken place. A copy of the html is saved and Quke will display it in your default browser. This can be useful to diagnose why something has failed, but there are times you may not want Quke to do this.

This change adds an option to the config that allows users of Quke to control whether failures get displayed or not.