Add a firefox driver settings helper to the config
Also we have added a helper method to the proxy config class which handles the logic of returning a correctly populated hash of settings used when registering the firefox driver with Capybara.

Admittedly this was mainly driven by CodeClimate kicking up a fuss, it at least means we can better test the different permutations that the settings hash can take based on the config values.