Use cucumber --fast-fail to stop on error
Prior to this change we handled telling Cucumber whether to quit or not should an error occur and `stop_on_error` is set.

However whilst looking into another change I spotted that Cucumber now has a `--fail-fast` option built in.

So the main thrust of this change is to use this rather than hand crank our own. This meant though that the logic to determine what arguments to set for cucumber which we pass to ParallelTests first needed to know about the flag.

I decided this meant the method which generates the arg now needed to access the configuration for multiple attributes, none of them actually to do with parallel tests.

Hence this change also refactors the `ParallelConfiguration` and `Configuration` to move the logic into `Configuration`. We also changed how `ParallelConfiguration` is initialized as it already needed access to config attributes, but now it also needed to access our new helper method for the cucumber arg.