Display progress when running non-parallel
By including the arguments `--serialize-stdout` and `--combine-stderr` to all calls to **ParallelTests** we were telling it not to output anything until the process has completed.

We believe this is needed when running using parallel else the output becomes confused. However when you are running without parallel this is no longer an issue, and users would expect to see something.