Fix issue with always running parallel (#93)
Found when using Quke that even if parallel was disabled it would always run parallel anyway even if `--single` had been passed in to ParallelTests.

After some testing found it was because we would still set `--group-by`. It must be that if this is set ParallelTests defaults back to attempting to run in parallel and ignores the `--single` argument.

Hence this change which now only sets `--group-by` if parallel is enabled in the config.