Migrate from helpers to webdrivers
In March 2019 the maintainer of [chromedriver-helper](https://github.com/flavorjones/chromedriver-helper) announced his intention to [end support for the gem](https://github.com/flavorjones/chromedriver-helper/issues/83).

He recommended everyone switch to using [webdrivers](https://github.com/titusfortner/webdrivers) instead.

So as of March 30 2019 chromedriver-helper has not been supported and the project archived. This change covers migrating to webdrivers. As webdrivers also handles the geckodriver for firefox as well as chromedriver, we're also able to drop [geckodriver-helper](https://github.com/DevicoSolutions/geckodriver-helper) in this change.

For reference this change does not just include switching the gems. webdrivers includes some rake tasks that provide information and functionality for the underlying drivers it installs. So we have added support and documentation to Quke to allow users to make those Rake tasks available in their projects.

It also manages the proxy for webdrivers. Essentially the first time you run the tests webdrivers makes a call to check if there has been an update for a driver. If there has it downloads and then installs the driver, before proceeding to run the tests (the default is to run the check once every 24 hours).

Because of this if you're project is sat behind a proxy you need to let webdrivers know.

Resolves #95