Remove support for phantomjs (#97)
[Phantomjs](https://github.com/ariya/phantomjs) is a headless WebKit scriptable with JavaScript. It was great in cases like running tests in CI as it didn't rely on additional dependencies like [xvfb](https://en.wikipedia.org/wiki/Xvfb) to allow a browser to run.

However the maintainer has stopped all development due to lack of support and contributions, possibly due to the main browsers now supporting headless options.

So with no ongoing support we don't think its feasible for Quke to continue to try and support it. Hence this change which removes it from the project, and sets Chrome as the new default.