Remove support for parallel tests
Back in PR #72 we added support for running the tests in parallel by implementing the [parallel_tests](https://github.com/grosser/parallel_tests) gem.

It worked, but we have then encountered issues passing arguments to cucumber such as in issue #94.

This seems to be because we're essentially calling one binary, which then calls another binary, and hoping that are arguments get passed through intact.

Because of this double binary issue it's also meant debugging of tests has stopped working.

Tl;DR we've never been very happy with this solution for running tests in parallel. On review we think its actually something we can implement directly in Quke itself which will hopefully nullify these issues.

So we're ripping parallel tests back out (sorry!) and then will try and re-implement them using our own simpler logic for handling the multiple threads.