Stop throwing error on cucumber test fails (#99)
If any tests fail when cucumber runs it throws a `SystemExit` with a failed status.

For no doubt valid reasons (at the time!) we handled this and then raised a `StandardError`. The result was that anyone running Quke that had any failing tests would see output that made it look like something was broken.

We don't believe we need to be doing this, and would rather have a clean exit after running cucumber.