Cleanup remnants from parallel_tests removal
When we removed [parallel_tests](https://github.com/grosser/parallel_tests) in PR #98 we weren't as thorough as we should have been.

Also by simply pasting in the old way of setting up the args for cucumber, we skipped taking advantage of some args we had previously added support for (`--format` and `--fail-fast`).

This this change is essentially a cleanup of the mess we left after removing parallel_tests.