Update args handling to reuse existing method
When we stripped out parallel tests we simply copy & pasted in the old way of parsing and generating the args we pass to Quke.

However as part of implenting parallel_test we'd implemented an improved version that included support for `--format` and `--fail-fast`.

This change updates the setting of the arguments to make use of the improved version and again add support for these Cucumber features.