Resolve issues and update spec structure
With the latest version of rubocop we were getting some errors with namespacing. It also highlgihted a deterministic sorting issue with how the `spec_helper.rb` was requiring support files. It was then we spotted that Quke's `spec_helper.rb` is really not that consistent with our other gems.

So this is a general fix and update change of minor things in the project.