Stop Quke from returning 0 exit status despite passing tests
https://eaflood.atlassian.net/browse/RUBY-1373

The default behaviour here is for Cucumber to exit as soon as a single test fails. We prefer to run the full suite each time, so these errors were being rescued.

However, this meant that if a test did fail, the exit status of the job would still be 0, which isn't what we want for our CI.

This PR changes the behaviour of Quke to record if a test fails, allow the test suite to proceed, and then throw the error at the end instead.